style_chek2

+++ tests/ford_bellman_test.cpp
@@ -6,10 +6,10 @@
  */
 
 #include <httplib.h>
+#include <vector>
+#include <random>
 #include <unordered_set>
 #include <nlohmann/json.hpp>
-#include <random>
-#include <vector>
 #include "test_core.hpp"
 
 static void SimpleTest(httplib::Client* cli);
@@ -28,7 +28,7 @@ static void SimpleTest(httplib::Client* cli) {
   tmp["graph_type"] = "WeightedOrientedGraph";
   tmp["weight_type"] = "int";
   tmp["vertices"] = std::vector<int>{0, 1, 2, 3, 4};
-  
+
   tmp["edges"][0]["from"] = 0;
   tmp["edges"][0]["to"] = 1;
   tmp["edges"][0]["weight"] = 1;
@@ -48,14 +48,14 @@ static void SimpleTest(httplib::Client* cli) {
   tmp["edges"][4]["from"] = 2;
   tmp["edges"][4]["to"] = 4;
   tmp["edges"][4]["weight"] = 2;
-  
+
   std::string input = tmp.dump();
   auto res = cli->Post("/FordBellman", input, "application/json");
 
   if (!res) {
     REQUIRE(false);
   }
-  
+
   nlohmann::json output = nlohmann::json::parse(res->body);
 
   std::vector<int> result = output.at("result");
@@ -87,10 +87,10 @@ static void RandomTest(httplib::Client* cli) {
   tmp["edges"][1]["to"] = 2;
   tmp["edges"][1]["weight"] = 2.0;
 
-  double w = weight1(gen1);   
+  double w = weight1(gen1);
   tmp["edges"][2]["from"] = 1;
   tmp["edges"][2]["to"] = 3;
-  tmp["edges"][2]["weight"] = (double)w;
+  tmp["edges"][2]["weight"] = w;
 
   tmp["edges"][3]["from"] = 4;
   tmp["edges"][3]["to"] = 3;
@@ -113,10 +113,10 @@ static void RandomTest(httplib::Client* cli) {
   std::vector<double> result = output.at("result");
 
   std::unordered_set<double> expected = {0, 1, 3, 6, 5};
-  
+
   std::unordered_set<double> resultSet;
-  
-  for(int id : result) {
+
+  for (int id : result) {
     resultSet.insert(id);
   }
   REQUIRE_EQUAL(expected, resultSet);
